requirements.txt
# Google API dependencies
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Additional dependencies
requests==2.31.0
urllib3==2.0.7
certifi==2023.7.22

# Development dependencies (optional)
# pytest==7.4.3
# black==23.9.1
# flake8==6.1.0